rename external metadata -wip
