+++ core/extdata/external_metadata_test.go
@@ -1,4 +1,4 @@
-package core
+package extdata
 
 import (
 	"context"
